refactor(types/query): reorganize, brand Condition, move ConditionNode
